+++ test/test_transfer.py
@@ -4,7 +4,7 @@ import sys
 flask_folder = \
     os.path.abspath(__file__).replace('\\', '/').rsplit('/flask_app', 1)[0]
 sys.path.append(flask_folder)
-from test.base import *
+from test.test_base import *
 from test.test_account import AccountTest
