+++ test/test_user.py
@@ -5,7 +5,7 @@ flask_folder = \
     os.path.abspath(__file__).replace('\\', '/').rsplit('/flask_app', 1)[0]
 sys.path.append(flask_folder)
 
-from test.base import *
+from test.test_base import *
 
 
 class UserTest(BaseTest):
